refactor: separate LoggerContext
