+++ apps/web/src/contexts/LoggerProvider.tsx
@@ -1,56 +1,11 @@
-import React, { createContext, useState, useEffect, ReactNode } from 'react';
+import React, { useState, useEffect, ReactNode } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-
-// ログレベルの定義
-/**
- * ログレベルの型定義
- * - debug: デバッグ情報
- * - info: 通常の操作・情報
- * - warn: 警告
- * - error: エラー
- */
-export type LogLevel = 'debug' | 'info' | 'warn' | 'error';
-
-// ログイベント構造の定義
-/**
- * ログイベント構造の定義
- * @property {number} timestamp - イベント発生時刻（UNIXエポックms）
- * @property {LogLevel} level - ログレベル
- * @property {string} action - アクション名
- * @property {Record<string, unknown>} [details] - 追加情報
- * @property {string} sessionId - セッションID
- */
-export interface LogEvent {
-  timestamp: number;
-  level: LogLevel;
-  action: string;
-  details?: Record<string, unknown>;
-  sessionId: string;
-}
-
-// ロガーコンテキストの型定義
-/**
- * LoggerContext で提供される関数・状態の型
- */
-export interface LoggerContextType {
-  /** ログ記録関数 */
-  log: (level: LogLevel, action: string, details?: Record<string, unknown>) => void;
-  /** 現在のログイベント配列 */
-  events: LogEvent[];
-  /** ログのクリア */
-  clearEvents: () => void;
-  /** セッションID */
-  sessionId: string;
-  /** ログを JSON 文字列でエクスポート */
-  exportLogs: () => string;
-}
-
-// コンテキストの作成
-/**
- * LoggerContext
- * ログ機能を提供するReact Context
- */
-export const LoggerContext = createContext<LoggerContextType | null>(null);
+import {
+  LoggerContext,
+  type LoggerContextType,
+  type LogEvent,
+  type LogLevel,
+} from './LoggerContext';
 
 // プロバイダーコンポーネント
 /**
