+++ apps/web/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { LoggerProvider } from './contexts/LoggerContext';
+import { LoggerProvider } from './contexts/LoggerProvider';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
